% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prior_spec.R
\name{add_prior}
\alias{add_prior}
\title{Add a prior to a PriorSpecification}
\usage{
add_prior(
  prior_spec,
  parameter,
  distribution,
  location = NA,
  scale = NA,
  lb = NA,
  ub = NA
)
}
\arguments{
\item{prior_spec}{A PriorSpecification object}

\item{parameter}{Name of the parameter}

\item{distribution}{Name of the distribution}

\item{location}{Location parameter}

\item{scale}{Scale parameter}

\item{lb}{Lower bound}

\item{ub}{Upper bound}
}
\value{
Updated PriorSpecification object
}
\description{
Add a prior to a PriorSpecification
}
