% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reporting.R
\name{add_step}
\alias{add_step}
\title{Add a step to an AnalysisWorkflow}
\usage{
add_step(workflow, name, function_name, args)
}
\arguments{
\item{workflow}{An AnalysisWorkflow object}

\item{name}{Name of the step}

\item{function_name}{Name of the function called}

\item{args}{Arguments passed to the function}
}
\value{
Updated AnalysisWorkflow object
}
\description{
Add a step to an AnalysisWorkflow
}
