% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/effect_mods.R
\name{assess_subgroup_credibility}
\alias{assess_subgroup_credibility}
\title{Assess credibility of subgroup effects}
\usage{
assess_subgroup_credibility(
  effect_results,
  criteria = c("pre_specified", "interaction", "consistency", "multiple_testing",
    "biological_plausibility")
)
}
\arguments{
\item{effect_results}{Results from effect modifier identification}

\item{criteria}{Vector of credibility criteria to assess}
}
\value{
Data frame with credibility assessment
}
\description{
Assess credibility of subgroup effects
}
