% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prior_spec.R
\name{auto_scale_priors}
\alias{auto_scale_priors}
\title{Auto-scale priors based on data}
\usage{
auto_scale_priors(prior_specs, formula, data)
}
\arguments{
\item{prior_specs}{PriorSpecification object}

\item{formula}{Model formula}

\item{data}{Data frame}
}
\value{
Updated PriorSpecification object
}
\description{
Auto-scale priors based on data
}
\keyword{internal}
