% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_fitting.R
\name{BayesianModel}
\alias{BayesianModel}
\title{BayesianModel class}
\usage{
BayesianModel(brms_model, model_spec, prior_spec = NULL, data_info = NULL)
}
\description{
S3 class for storing fitted Bayesian models and metadata
}
