% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_spec.R
\name{build_formula}
\alias{build_formula}
\title{Build a brms formula from ModelSpecification}
\usage{
build_formula(spec)
}
\arguments{
\item{spec}{A ModelSpecification object}
}
\value{
A brms formula object
}
\description{
Build a brms formula from ModelSpecification
}
