% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reporting.R
\name{capture_workflow}
\alias{capture_workflow}
\title{Capture workflow steps}
\usage{
capture_workflow(analysis_steps, save_path = "workflow.json")
}
\arguments{
\item{analysis_steps}{List of analysis step specifications}

\item{save_path}{Path to save the workflow file}
}
\value{
An AnalysisWorkflow object
}
\description{
Capture workflow steps
}
