% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnostics.R
\name{check_residuals}
\alias{check_residuals}
\title{Check residuals for a Bayesian model}
\usage{
check_residuals(
  model,
  type = c("default", "pearson", "response"),
  ndraws = 100,
  plot = TRUE
)
}
\arguments{
\item{model}{A BayesianModel object}

\item{type}{Type of residuals (default, pearson, response)}

\item{ndraws}{Number of posterior draws to use}

\item{plot}{Logical indicating whether to create a plot}
}
\value{
Data frame of residuals or a plot if requested
}
\description{
Check residuals for a Bayesian model
}
