% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prior_spec.R
\name{construct_parameter_name}
\alias{construct_parameter_name}
\title{Construct a parameter name from brms prior components}
\usage{
construct_parameter_name(row)
}
\arguments{
\item{row}{A row from brms get_prior output}
}
\value{
Character string with parameter name
}
\description{
Construct a parameter name from brms prior components
}
\keyword{internal}
