% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{create_log_message}
\alias{create_log_message}
\title{Create a standardized log message}
\usage{
create_log_message(msg, level = "info", timestamp = TRUE)
}
\arguments{
\item{msg}{The message to log}

\item{level}{Character string of log level (info, warning, error, debug)}

\item{timestamp}{Logical indicating whether to include timestamp}
}
\value{
Character string with formatted log message
}
\description{
Create a standardized log message
}
\keyword{internal}
