% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_prep.R
\name{create_trial_dataset}
\alias{create_trial_dataset}
\title{Create standardized trial dataset}
\usage{
create_trial_dataset(
  data,
  id_var,
  treatment_var,
  outcome_var,
  covariates = NULL,
  time_var = NULL
)
}
\arguments{
\item{data}{Data frame with trial data}

\item{id_var}{Name of ID variable}

\item{treatment_var}{Name of treatment variable}

\item{outcome_var}{Name of outcome variable}

\item{covariates}{Vector of covariate variable names}

\item{time_var}{Name of time variable (optional)}
}
\value{
Standardized trial dataset
}
\description{
Create standardized trial dataset
}
