% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prior_spec.R
\name{draw_prior_samples}
\alias{draw_prior_samples}
\title{Draw samples from a prior distribution}
\usage{
draw_prior_samples(prior_row, n = 1000)
}
\arguments{
\item{prior_row}{Row from priors data frame}

\item{n}{Number of samples to draw}
}
\value{
Vector of samples
}
\description{
Draw samples from a prior distribution
}
\keyword{internal}
