% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_fitting.R
\name{extract_posterior}
\alias{extract_posterior}
\title{Extract posterior samples from a model}
\usage{
extract_posterior(model, variable, transform = NULL)
}
\arguments{
\item{model}{A BayesianModel object}

\item{variable}{Name of the variable to extract}

\item{transform}{Function to transform samples (optional)}
}
\value{
Data frame of posterior samples
}
\description{
Extract posterior samples from a model
}
