% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_fitting.R
\name{fit_model}
\alias{fit_model}
\title{Fit a Bayesian model using brms}
\usage{
fit_model(
  model_spec,
  data,
  chains = 4,
  cores = parallel::detectCores(),
  iter = 2000,
  seed = NULL,
  save_model = TRUE,
  control = list(adapt_delta = 0.9)
)
}
\arguments{
\item{model_spec}{A ModelSpecification object}

\item{data}{Data frame containing the data}

\item{chains}{Number of MCMC chains}

\item{cores}{Number of cores to use}

\item{iter}{Number of iterations per chain}

\item{seed}{Random seed for reproducibility}

\item{save_model}{Logical indicating whether to save model object}

\item{control}{List of control parameters for Stan}
}
\value{
A BayesianModel object
}
\description{
Fit a Bayesian model using brms
}
