% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_fitting.R
\name{fit_models}
\alias{fit_models}
\title{Fit multiple models}
\usage{
fit_models(model_specs, data, parallel = TRUE, ...)
}
\arguments{
\item{model_specs}{List of ModelSpecification objects}

\item{data}{Data frame containing the data}

\item{parallel}{Logical indicating whether to use parallel processing}

\item{...}{Additional arguments passed to fit_model}
}
\value{
List of BayesianModel objects
}
\description{
Fit multiple models
}
