% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parallel.R
\name{fit_models_parallel}
\alias{fit_models_parallel}
\title{Fit Models in Parallel}
\usage{
fit_models_parallel(
  model_specs,
  data,
  priors_list = NULL,
  fit_args = list(),
  progress = TRUE
)
}
\arguments{
\item{model_specs}{List of model specifications}

\item{data}{Data frame for fitting}

\item{priors_list}{List of prior specifications}

\item{fit_args}{Additional arguments for model fitting}

\item{progress}{Whether to show progress}
}
\value{
List of fitted models
}
\description{
Internal function to fit multiple models in parallel using the configured backend.
}
\keyword{internal}
