% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{format_value}
\alias{format_value}
\title{Format a value for printing based on its type}
\usage{
format_value(x)
}
\arguments{
\item{x}{The value to format}
}
\value{
Character string representation of the value
}
\description{
Format a value for printing based on its type
}
\keyword{internal}
