% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_option}
\alias{get_option}
\title{Get package option with fallback to default}
\usage{
get_option(option_name, default = NULL)
}
\arguments{
\item{option_name}{Name of the option to retrieve}

\item{default}{Default value if option is not set}
}
\value{
The option value or default
}
\description{
Get package option with fallback to default
}
\keyword{internal}
