% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_session_info}
\alias{get_session_info}
\title{Get session info including package versions}
\usage{
get_session_info()
}
\value{
List containing session information
}
\description{
Get session info including package versions
}
