% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reporting.R
\name{get_template_content}
\alias{get_template_content}
\title{Get content for standard templates}
\usage{
get_template_content(template_name)
}
\arguments{
\item{template_name}{Name of the template}
}
\value{
Template content as character string
}
\description{
Get content for standard templates
}
\keyword{internal}
