% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_synthetic_data.R
\name{get_variable_descriptions}
\alias{get_variable_descriptions}
\title{Get Variable Descriptions for Synthetic Data}
\usage{
get_variable_descriptions(data)
}
\arguments{
\item{data}{Synthetic data generated by generate_synthetic_data()}
}
\value{
A tibble with variable descriptions
}
\description{
Provides descriptions of variables in synthetic clinical trial data.
}
