% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_working_dir}
\alias{get_working_dir}
\title{Get current working directory and format as absolute path}
\usage{
get_working_dir()
}
\value{
Character string of absolute path to current working directory
}
\description{
Get current working directory and format as absolute path
}
\keyword{internal}
