% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_prep.R
\name{harmonize_trials}
\alias{harmonize_trials}
\title{Harmonize variables across trials}
\usage{
harmonize_trials(
  trial_list,
  mapping_file = NULL,
  auto_detect = TRUE,
  standard = c("none", "sdtm", "adam")
)
}
\arguments{
\item{trial_list}{List of trial data frames}

\item{mapping_file}{Path to variable mapping CSV file (optional)}

\item{auto_detect}{Logical indicating whether to attempt automatic variable detection}

\item{standard}{Data standard to use (none, sdtm, adam)}
}
\value{
List of harmonized trial data frames
}
\description{
Harmonize variables across trials
}
