% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/effect_mods.R
\name{identify_bart_effects}
\alias{identify_bart_effects}
\title{Identify effect modifiers using Bayesian Additive Regression Trees}
\usage{
identify_bart_effects(
  model,
  data,
  treatment_var,
  candidate_modifiers,
  credibility_threshold = 0.95,
  n_trees = 200,
  ...
)
}
\arguments{
\item{model}{A BayesianModel object or formula}

\item{data}{Data frame containing the data}

\item{treatment_var}{Name of the treatment variable}

\item{candidate_modifiers}{Vector of candidate effect modifier variables}

\item{credibility_threshold}{Threshold for credibility of effect modification}

\item{n_trees}{Number of trees for BART model}

\item{...}{Additional arguments passed to BART functions}
}
\value{
List of effect modifier results
}
\description{
Identify effect modifiers using Bayesian Additive Regression Trees
}
\keyword{internal}
