% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/effect_mods.R
\name{identify_continuous_effects}
\alias{identify_continuous_effects}
\title{Identify effect modifiers using continuous interaction modeling}
\usage{
identify_continuous_effects(
  model,
  data,
  treatment_var,
  candidate_modifiers,
  credibility_threshold = 0.95,
  use_splines = TRUE,
  spline_df = 3,
  ...
)
}
\arguments{
\item{model}{A BayesianModel object or formula}

\item{data}{Data frame containing the data}

\item{treatment_var}{Name of the treatment variable}

\item{candidate_modifiers}{Vector of candidate effect modifier variables}

\item{credibility_threshold}{Threshold for credibility of effect modification}

\item{use_splines}{Logical indicating whether to use splines for continuous modifiers}

\item{spline_df}{Degrees of freedom for splines}

\item{...}{Additional arguments passed to fit_model}
}
\value{
List of effect modifier results
}
\description{
Identify effect modifiers using continuous interaction modeling
}
\keyword{internal}
