% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/effect_mods.R
\name{identify_effect_modifiers}
\alias{identify_effect_modifiers}
\title{Identify effect modifiers}
\usage{
identify_effect_modifiers(
  model,
  data,
  treatment_var,
  candidate_modifiers,
  method = c("interaction", "bart", "continuous"),
  credibility_threshold = 0.95,
  ...
)
}
\arguments{
\item{model}{A BayesianModel object or formula}

\item{data}{Data frame containing the data}

\item{treatment_var}{Name of the treatment variable}

\item{candidate_modifiers}{Vector of candidate effect modifier variables}

\item{method}{Method for identifying effect modifiers}

\item{credibility_threshold}{Threshold for credibility of effect modification}

\item{...}{Additional arguments passed to specific methods}
}
\value{
List of effect modifier results
}
\description{
Identify effect modifiers
}
