% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{is.formula}
\alias{is.formula}
\title{Check if a variable is a formula}
\usage{
is.formula(x)
}
\arguments{
\item{x}{Object to check}
}
\value{
Logical indicating if x is a formula
}
\description{
Check if a variable is a formula
}
\keyword{internal}
