% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{is_class}
\alias{is_class}
\title{Check if object is of a specific class}
\usage{
is_class(x, class_name)
}
\arguments{
\item{x}{Object to check}

\item{class_name}{Name of the class to check for}
}
\value{
Logical indicating if object inherits from class_name
}
\description{
Check if object is of a specific class
}
\keyword{internal}
