% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_prep.R
\name{load_clinical_trial}
\alias{load_clinical_trial}
\title{Load clinical trial data from a CSV file}
\usage{
load_clinical_trial(file_path, data_format = c("default", "sdtm", "adam"), ...)
}
\arguments{
\item{file_path}{Path to CSV file}

\item{data_format}{Format of the data (default, sdtm, adam)}

\item{...}{Additional arguments passed to read.csv}
}
\value{
A data frame containing clinical trial data
}
\description{
Load clinical trial data from a CSV file
}
