% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_prep.R
\name{load_trials}
\alias{load_trials}
\title{Load multiple clinical trials}
\usage{
load_trials(
  file_paths,
  trial_names = NULL,
  data_format = c("default", "sdtm", "adam"),
  ...
)
}
\arguments{
\item{file_paths}{Vector of paths to CSV files}

\item{trial_names}{Vector of names for the trials (optional)}

\item{data_format}{Format of the data (default, sdtm, adam)}

\item{...}{Additional arguments passed to read.csv}
}
\value{
A list of data frames containing clinical trial data
}
\description{
Load multiple clinical trials
}
