% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reporting.R
\name{load_workflow}
\alias{load_workflow}
\title{Load AnalysisWorkflow from a file}
\usage{
load_workflow(file_path)
}
\arguments{
\item{file_path}{Path to the workflow file}
}
\value{
An AnalysisWorkflow object
}
\description{
Load AnalysisWorkflow from a file
}
