% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{log_message}
\alias{log_message}
\title{Log a message to console or file}
\usage{
log_message(msg, level = "info", file = NULL, console = TRUE)
}
\arguments{
\item{msg}{The message to log}

\item{level}{Character string of log level (info, warning, error, debug)}

\item{file}{Optional file path to write log to}

\item{console}{Logical indicating whether to print to console}
}
\value{
Invisibly returns the formatted log message
}
\description{
Log a message to console or file
}
