% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnostics.R
\name{model_fit_stats}
\alias{model_fit_stats}
\title{Extract model fit statistics}
\usage{
model_fit_stats(model, type = c("loo", "waic"))
}
\arguments{
\item{model}{A BayesianModel object}

\item{type}{Type of fit statistic to compute (loo, waic)}
}
\value{
Object containing model fit statistics
}
\description{
Extract model fit statistics
}
