% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_spec.R
\name{ModelSpecification}
\alias{ModelSpecification}
\title{ModelSpecification class}
\usage{
ModelSpecification(
  model_name = NULL,
  outcome = NULL,
  predictors = NULL,
  random_effects = NULL,
  family = "gaussian",
  link = "identity",
  additional_args = list()
)
}
\description{
S3 class for storing and validating model specifications
}
