% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prior_spec.R
\name{parse_parameter_name}
\alias{parse_parameter_name}
\title{Parse parameter name into components}
\usage{
parse_parameter_name(param_name)
}
\arguments{
\item{param_name}{Name of the parameter}
}
\value{
List with class, coef, and group components
}
\description{
Parse parameter name into components
}
\keyword{internal}
