% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/effect_mods.R, R/visualization.R
\name{plot_interaction}
\alias{plot_interaction}
\title{Plot effect modifier results}
\usage{
plot_interaction(
  model,
  treatment_var,
  modifier_var,
  data = NULL,
  type = c("continuous", "categorical")
)

plot_interaction(
  model,
  treatment_var,
  modifier_var,
  data = NULL,
  type = c("continuous", "categorical")
)
}
\arguments{
\item{model}{A BayesianModel object}

\item{treatment_var}{Name of the treatment variable}

\item{modifier_var}{Name of the effect modifier variable}

\item{data}{Data frame (optional)}

\item{type}{Type of interaction plot}

\item{effect_results}{Results from effect modifier identification}

\item{modifier}{Name of the modifier to plot}

\item{plot_type}{Type of plot to create}
}
\value{
A ggplot object

A ggplot object
}
\description{
Plot effect modifier results

Plot interaction effects
}
