% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{plot_model_comparison}
\alias{plot_model_comparison}
\title{Plot model comparison}
\usage{
plot_model_comparison(
  models,
  parameter,
  plot_type = c("forest", "ridgeline", "intervals")
)
}
\arguments{
\item{models}{List of BayesianModel objects}

\item{parameter}{Name of the parameter to compare}

\item{plot_type}{Type of comparison plot}
}
\value{
A ggplot object
}
\description{
Plot model comparison
}
