% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{plot_model_fit}
\alias{plot_model_fit}
\title{Plot model fit}
\usage{
plot_model_fit(
  model,
  type = c("predicted_vs_observed", "residuals", "pp_check"),
  n_draws = 50
)
}
\arguments{
\item{model}{A BayesianModel object}

\item{type}{Type of fit plot}

\item{n_draws}{Number of posterior draws to plot}
}
\value{
A ggplot object
}
\description{
Plot model fit
}
