% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{plot_posterior}
\alias{plot_posterior}
\title{Plot posterior distributions}
\usage{
plot_posterior(
  posterior_data,
  plot_type = c("halfeye", "interval", "dots"),
  facet_by = NULL,
  color_by = NULL,
  theme = "publication"
)
}
\arguments{
\item{posterior_data}{Data frame or matrix of posterior samples}

\item{plot_type}{Type of posterior plot to create}

\item{facet_by}{Variable to facet plots by}

\item{color_by}{Variable to color plots by}

\item{theme}{Plot theme to use}
}
\value{
A ggplot object
}
\description{
Plot posterior distributions
}
