% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prior_spec.R
\name{plot_priors}
\alias{plot_priors}
\title{Visualize prior distributions}
\usage{
plot_priors(prior_specs, parameters = NULL, n_samples = 1000)
}
\arguments{
\item{prior_specs}{PriorSpecification object}

\item{parameters}{Vector of parameter names to visualize (NULL for all)}

\item{n_samples}{Number of samples to draw from each prior}
}
\value{
A ggplot object
}
\description{
Visualize prior distributions
}
