% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sensitivity.R
\name{plot_sensitivity}
\alias{plot_sensitivity}
\title{Plot sensitivity analysis results}
\usage{
plot_sensitivity(
  sensitivity_results,
  parameter,
  plot_type = c("forest", "ridgeline", "scatter")
)
}
\arguments{
\item{sensitivity_results}{Results from run_sensitivity_analyses}

\item{parameter}{Name of the parameter to plot}

\item{plot_type}{Type of plot to create (forest, ridgeline, scatter)}
}
\value{
A ggplot object
}
\description{
Plot sensitivity analysis results
}
