% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnostics.R
\name{posterior_predictive_pvalue}
\alias{posterior_predictive_pvalue}
\title{Compute posterior predictive p-value}
\usage{
posterior_predictive_pvalue(model, test_statistic = NULL, nsamples = 1000)
}
\arguments{
\item{model}{A BayesianModel object}

\item{test_statistic}{Function computing test statistic from observed and replicated data}

\item{nsamples}{Number of posterior samples to use}
}
\value{
Posterior predictive p-value
}
\description{
Compute posterior predictive p-value
}
