% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_prep.R
\name{preprocess_data}
\alias{preprocess_data}
\title{Preprocess data based on specifications}
\usage{
preprocess_data(
  data,
  specs = NULL,
  steps = c("transform", "recode", "convert", "center", "scale")
)
}
\arguments{
\item{data}{Data frame to preprocess}

\item{specs}{Data frame or path to CSV file with covariate specifications}

\item{steps}{Character vector of preprocessing steps to perform}
}
\value{
Preprocessed data frame
}
\description{
Preprocess data based on specifications
}
