% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_fitting.R
\name{print.BayesianModel}
\alias{print.BayesianModel}
\title{Print method for BayesianModel}
\usage{
\method{print}{BayesianModel}(x, ...)
}
\arguments{
\item{x}{A BayesianModel object}

\item{...}{Additional arguments passed to print}
}
\value{
x invisibly
}
\description{
Print method for BayesianModel
}
