% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enhanced_diagnostics.R
\name{print.bayestrials_diagnostics}
\alias{print.bayestrials_diagnostics}
\title{Print Diagnostic Summary}
\usage{
\method{print}{bayestrials_diagnostics}(x, ...)
}
\arguments{
\item{x}{A bayestrials_diagnostics object}

\item{...}{Additional arguments (ignored)}
}
\value{
x invisibly
}
\description{
Print Diagnostic Summary
}
