% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parallel.R
\name{print.bayestrials_parallel_config}
\alias{print.bayestrials_parallel_config}
\title{Print Parallel Configuration}
\usage{
\method{print}{bayestrials_parallel_config}(x, ...)
}
\arguments{
\item{x}{A bayestrials_parallel_config object}

\item{...}{Additional arguments (ignored)}
}
\value{
x invisibly
}
\description{
Print Parallel Configuration
}
