% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_spec.R
\name{print.ModelSpecification}
\alias{print.ModelSpecification}
\title{Print method for ModelSpecification}
\usage{
\method{print}{ModelSpecification}(x, ...)
}
\arguments{
\item{x}{A ModelSpecification object}

\item{...}{Additional arguments passed to print}
}
\value{
x invisibly
}
\description{
Print method for ModelSpecification
}
