% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prior_spec.R
\name{print.PriorSpecification}
\alias{print.PriorSpecification}
\title{Print method for PriorSpecification}
\usage{
\method{print}{PriorSpecification}(x, ...)
}
\arguments{
\item{x}{A PriorSpecification object}

\item{...}{Additional arguments passed to print}
}
\value{
x invisibly
}
\description{
Print method for PriorSpecification
}
