% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_prep.R
\name{process_adam_data}
\alias{process_adam_data}
\title{Process ADaM format data}
\usage{
process_adam_data(data)
}
\arguments{
\item{data}{Data frame in ADaM format}
}
\value{
Processed data frame
}
\description{
Process ADaM format data
}
\keyword{internal}
