% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_prep.R
\name{process_sdtm_data}
\alias{process_sdtm_data}
\title{Process SDTM format data}
\usage{
process_sdtm_data(data)
}
\arguments{
\item{data}{Data frame in SDTM format}
}
\value{
Processed data frame
}
\description{
Process SDTM format data
}
\keyword{internal}
