% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prior_spec.R
\name{read_prior_spec}
\alias{read_prior_spec}
\title{Read prior specifications from CSV}
\usage{
read_prior_spec(file_path)
}
\arguments{
\item{file_path}{Path to CSV file with prior specifications}
}
\value{
A PriorSpecification object
}
\description{
Read prior specifications from CSV
}
