% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prior_spec.R
\name{remove_prior}
\alias{remove_prior}
\title{Remove a prior from a PriorSpecification}
\usage{
remove_prior(prior_spec, parameter)
}
\arguments{
\item{prior_spec}{A PriorSpecification object}

\item{parameter}{Name of the parameter to remove}
}
\value{
Updated PriorSpecification object
}
\description{
Remove a prior from a PriorSpecification
}
