% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_prep.R
\name{rename_variables}
\alias{rename_variables}
\title{Rename variables in a data frame}
\usage{
rename_variables(data, mapping)
}
\arguments{
\item{data}{Data frame to rename variables in}

\item{mapping}{Named vector of new names, where names are old variable names}
}
\value{
Data frame with renamed variables
}
\description{
Rename variables in a data frame
}
\keyword{internal}
