% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reporting.R
\name{restore_analysis}
\alias{restore_analysis}
\title{Restore analysis from workflow}
\usage{
restore_analysis(workflow_path, verify_packages = TRUE)
}
\arguments{
\item{workflow_path}{Path to workflow file}

\item{verify_packages}{Logical indicating whether to verify package versions}
}
\value{
Results of the last step in the workflow
}
\description{
Restore analysis from workflow
}
