% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_spec.R
\name{save_model_specs}
\alias{save_model_specs}
\title{Save model specifications to CSV}
\usage{
save_model_specs(
  model_specs,
  model_file,
  prior_file = NULL,
  covariate_file = NULL
)
}
\arguments{
\item{model_specs}{List of ModelSpecification objects}

\item{model_file}{Path to save model specifications}

\item{prior_file}{Path to save prior specifications (optional)}

\item{covariate_file}{Path to save covariate specifications (optional)}
}
\value{
Invisibly returns TRUE if successful
}
\description{
Save model specifications to CSV
}
