% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prior_spec.R
\name{save_prior_spec}
\alias{save_prior_spec}
\title{Save prior specifications to CSV}
\usage{
save_prior_spec(prior_spec, file_path)
}
\arguments{
\item{prior_spec}{A PriorSpecification object}

\item{file_path}{Path to save prior specifications}
}
\value{
Invisibly returns TRUE if successful
}
\description{
Save prior specifications to CSV
}
