% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reporting.R
\name{save_workflow}
\alias{save_workflow}
\title{Save AnalysisWorkflow to a file}
\usage{
save_workflow(workflow, file_path)
}
\arguments{
\item{workflow}{An AnalysisWorkflow object}

\item{file_path}{Path to save the workflow}
}
\value{
Invisibly returns TRUE if successful
}
\description{
Save AnalysisWorkflow to a file
}
