% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{set_option}
\alias{set_option}
\title{Set package option}
\usage{
set_option(option_name, value)
}
\arguments{
\item{option_name}{Name of the option to set}

\item{value}{Value to set for the option}
}
\value{
Invisibly returns the previous option value
}
\description{
Set package option
}
