% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{set_seed}
\alias{set_seed}
\title{Set a global random seed in a reproducible way}
\usage{
set_seed(seed = NULL)
}
\arguments{
\item{seed}{Integer seed for random number generation}
}
\value{
Invisibly returns the seed used
}
\description{
Set a global random seed in a reproducible way
}
