% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prior_spec.R
\name{specify_priors}
\alias{specify_priors}
\title{Specify priors for a model}
\usage{
specify_priors(
  model_formula,
  prior_specs = NULL,
  prior_type = c("default", "weakly_informative", "skeptical", "optimistic", "reference"),
  auto_scale = TRUE,
  data = NULL
)
}
\arguments{
\item{model_formula}{Model formula or ModelSpecification}

\item{prior_specs}{PriorSpecification object or NULL}

\item{prior_type}{Type of priors to use (default, weakly_informative, skeptical, optimistic, reference)}

\item{auto_scale}{Logical indicating whether to automatically scale priors}

\item{data}{Optional data for automatic scaling}
}
\value{
A PriorSpecification object
}
\description{
Specify priors for a model
}
