% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{str_to_formula}
\alias{str_to_formula}
\title{Convert a string to a formula}
\usage{
str_to_formula(formula_str)
}
\arguments{
\item{formula_str}{String representing a formula}
}
\value{
A formula object
}
\description{
Convert a string to a formula
}
\keyword{internal}
