% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/effect_mods.R
\name{summarize_effect_modifiers}
\alias{summarize_effect_modifiers}
\title{Summarize effect modifier results}
\usage{
summarize_effect_modifiers(effect_results)
}
\arguments{
\item{effect_results}{Results from effect modifier identification}
}
\value{
Text summary of effect modifiers
}
\description{
Summarize effect modifier results
}
\keyword{internal}
