% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prior_spec.R
\name{to_brms_prior}
\alias{to_brms_prior}
\title{Convert PriorSpecification to brms priors}
\usage{
to_brms_prior(prior_spec, formula = NULL)
}
\arguments{
\item{prior_spec}{A PriorSpecification object}

\item{formula}{A model formula or ModelSpecification}
}
\value{
A brms prior object
}
\description{
Convert PriorSpecification to brms priors
}
