% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_spec.R
\name{update_model_spec}
\alias{update_model_spec}
\title{Update existing model specification}
\usage{
update_model_spec(model_spec, ..., priors = NULL)
}
\arguments{
\item{model_spec}{A ModelSpecification object}

\item{...}{Additional parameters to update}

\item{priors}{New prior specifications (optional)}
}
\value{
Updated ModelSpecification object
}
\description{
Update existing model specification
}
