% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_spec.R
\name{validate_model_spec}
\alias{validate_model_spec}
\title{Validate a ModelSpecification object}
\usage{
validate_model_spec(spec)
}
\arguments{
\item{spec}{A ModelSpecification object}
}
\value{
Logical indicating if the specification is valid
}
\description{
Validate a ModelSpecification object
}
